Move webapp.common.generate_dn_hash() to a new file, x509.py
This lets you import webapp.common without having "asn1" installed,
which is good because get-scitokens-mapfile uses webapp.common and
asn1 isn't available as an RPM on EL7.
